teleops_backend/requirements/base.txt
# Core Django Framework
Django==4.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
django-cors-headers==4.6.0
django-filter==24.3
django-ratelimit==4.1.0

# File Storage & Media
azure-storage-blob==12.19.0
Pillow==10.2.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-yasg==1.21.8

# Data Validation & Serialization
marshmallow==3.20.1
django-dirtyfields==1.9.5

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Logging & Monitoring
structlog==23.2.0
django-health-check==3.18.1

# Database Migrations
django-migration-linter==3.0.0

# Performance & Optimization
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0 